+++ opentelemetry-instrumentation-tower/examples/axum-http-service/src/main.rs
@@ -57,7 +57,7 @@ async fn main() {
 
     global::set_meter_provider(meter_provider);
     
-    let otel_metrics_service_layer = otel_tower_metrics::HTTPMetricsLayerBuilder::builder()
+    let otel_metrics_service_layer = otel_tower_metrics::HTTPLayerBuilder::builder()
         .build()
         .unwrap();